feat: extract reusable components into separate component library
- Create @soroban-scanner/ui-components package
- Extract DraggableList component with drag-drop functionality
- Extract LazyImage component with lazy loading and WebP support
- Extract useWebSocket hook with reconnection logic
- Extract cache utility for memory and persistent storage
- Add comprehensive documentation and examples
- Update main project to use component library via re-exports
- Configure build system with Rollup and TypeScript
- Add ESLint and Jest configuration for component library
