+++ component-library/src/components/index.ts
@@ -1,0 +1,5 @@
+export { DraggableList } from './DraggableList';
+export { LazyImage } from './LazyImage';
+
+export type { DraggableListProps } from './DraggableList';
+export type { LazyImageProps } from './LazyImage';
