+++ component-library/src/hooks/index.ts
@@ -1,0 +1,4 @@
+export { useDragDrop } from './useDragDrop';
+export { useWebSocket } from './useWebSocket';
+
+export type { UseWebSocketOptions } from './useWebSocket';
